Improvements to flight deer
Tweak numbers to make 'walking' around on the platform possible.

Add basic blueprint of upgrade system (rockets are the only implemented
upgrade)

Add support for areas detecting the deer_missile: (can be used for
boosters etc on the ramp later)
LaunchZone area which changes your flight_state to 'FLIGHT' when you
exit.

Add deer animation blend tree so the reindeer runs when you move
forward/back.

Integrate UiPlay menu with flight_main scene so we can tell it what
flight_state we are in and exit when it's over