Deer Launch Sequence
closes #22

Also add icicle clusters around the ramp that you run over when you
launch.

For the final deer launch sequence we should have user able to position
the deer using left and right inputs... with a countdown that forces
them to go.

also tweaked the launch so it eliminates angular momentum when you leave
the ramp.. and made a base `deer_area.gd` that can be used for any area
that wants to affect the deer...